Extracted embedded cassandra to a separate module
