+++ cassandra-diagnostics-connector21/pom.xml
@@ -11,9 +11,9 @@
 
 	<properties>
 		<version.bytebuddy>1.4.9</version.bytebuddy>
-		<version.cassandra>2.1.14</version.cassandra>
+		<version.cassandra>2.1.17</version.cassandra>
 		<version.spring_instrument>4.2.6.RELEASE</version.spring_instrument>
-		<version.cassandra.driver>2.1.10.2</version.cassandra.driver>
+		<version.cassandra.driver>2.1.10.3</version.cassandra.driver>
 		<version.powermock>1.6.5</version.powermock>
 	</properties>
 
@@ -21,6 +21,11 @@
 		<dependency>
 			<groupId>io.smartcat</groupId>
 			<artifactId>cassandra-diagnostics-commons</artifactId>
+			<version>${project.parent.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>io.smartcat</groupId>
+			<artifactId>cassandra-diagnostics-embedded-cassandra</artifactId>
 			<version>${project.parent.version}</version>
 		</dependency>
 		<dependency>
