+++ cassandra-diagnostics-connector30/pom.xml
@@ -13,9 +13,9 @@
 
     <properties>
         <version.bytebuddy>1.4.9</version.bytebuddy>
-        <version.cassandra>3.0.8</version.cassandra>
+        <version.cassandra>3.0.12</version.cassandra>
         <version.spring_instrument>4.2.6.RELEASE</version.spring_instrument>
-        <version.cassandra.driver>3.0.1</version.cassandra.driver>
+        <version.cassandra.driver>3.1.4</version.cassandra.driver>
         <version.powermock>1.6.5</version.powermock>
     </properties>
 
@@ -23,6 +23,11 @@
         <dependency>
             <groupId>io.smartcat</groupId>
             <artifactId>cassandra-diagnostics-commons</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.smartcat</groupId>
+            <artifactId>cassandra-diagnostics-embedded-cassandra</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
         <dependency>
