+++ cassandra-diagnostics-embedded-cassandra/pom.xml
@@ -1,0 +1,39 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>io.smartcat</groupId>
+        <artifactId>cassandra-diagnostics</artifactId>
+        <version>1.3.9-SNPASHOT</version>
+    </parent>
+    <artifactId>cassandra-diagnostics-embedded-cassandra</artifactId>
+    <packaging>jar</packaging>
+
+    <name>cassandra-diagnostics-embedded-cassandra</name>
+
+    <properties>
+        <version.cassandra>2.1.17</version.cassandra>
+        <version.cassandra.driver>2.1.10.3</version.cassandra.driver>
+    </properties>
+
+    <dependencies>
+        <dependency>
+            <groupId>org.apache.cassandra</groupId>
+            <artifactId>cassandra-all</artifactId>
+            <version>${version.cassandra}</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.datastax.cassandra</groupId>
+            <artifactId>cassandra-driver-core</artifactId>
+            <version>${version.cassandra.driver}</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>3.8.1</version>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
+</project>
